+++ src/Page/Admin/Login/AdminLoadingAndGetUser.tsx
@@ -1,6 +1,7 @@
 import React, { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import { useRecoilState } from "recoil";
+
 import Loading from "../../../Components/Loading";
 import { MeQuery, useMeQuery } from "../../../lib/generated/graphql";
 import graphqlReqeustClient from "../../../lib/graphqlRequestClient";
