+++ src/Page/Admin/Login/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useLogin } from "./useLogin";
