+++ src/Page/Admin/Login/index.ts
@@ -1,0 +1,3 @@
+export type { IUserProps, ErrorState } from "./interface";
+export { default as AdminLoadingAndGetUser } from "./AdminLoadingAndGetUser";
+export { default as AdminLogin } from "./AdminLogin";
