+++ src/Page/Admin/Store/AdminStoreList.tsx
@@ -1,6 +1,5 @@
-import styled from "styled-components";
 import PageHeaderMessage from "../../../Components/PageHeader";
-import ButtonDefaultStyle from "../../../Components/Buttons/ButtonDefault";
+
 import { MdAddCircle } from "react-icons/md";
 import { Link, useNavigate } from "react-router-dom";
 import { useRecoilState, useRecoilValue } from "recoil";
@@ -12,7 +11,7 @@ import {
 } from "../../../lib/generated/graphql";
 import graphqlReqeustClient from "../../../lib/graphqlRequestClient";
 import { storesState, storeStateProps } from "../../../lib/state/storeState";
-import { Headline3 } from "../../../lib/styles/mixin";
+
 import { useEffect, useState } from "react";
 import Modal from "../../../Components/Modals/Modal";
 import IsOpenModalChildren from "../Modal/IsOpenModalChildren";
@@ -20,119 +19,14 @@ import { useQueryClient } from "react-query";
 import { MdDelete, MdCreate } from "react-icons/md";
 import useModalHook from "../../../lib/utils/customHooks/useModalHook";
 import ToggleButton from "../../../Components/Buttons/ToggleButton";
-
-const Wrapper = styled.div``;
-const Header = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  ${({ theme }) => theme.device.tablet} {
-    margin-bottom: 1rem;
-    flex-wrap: wrap;
-  }
-  ${({ theme }) => theme.device.mobile} {
-    margin-bottom: 1rem;
-  }
-`;
-
-const AddStoreButton = styled(ButtonDefaultStyle)`
-  display: flex;
-  align-items: center;
-  color: ${(props) => props.theme.color.fontColorBlack};
-  background-color: unset;
-  a {
-    text-decoration: none;
-    padding-bottom: 0.4rem;
-    color: ${(props) => props.theme.color.fontColorBlack};
-  }
-  span {
-    margin-left: 0.5rem;
-  }
-  ${({ theme }) => theme.device.tablet} {
-    padding: 0;
-  }
-  ${({ theme }) => theme.device.mobile} {
-    align-self: flex-end;
-    font-size: 2rem;
-  }
-`;
-
-const StoreList = styled.ul`
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  grid-gap: 1rem;
-  grid-auto-rows: minmax(15rem, 15rem);
-  ${({ theme }) => theme.device.tablet} {
-    grid-template-columns: auto;
-  }
-`;
-
-const Item = styled.li`
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  height: inherit;
-  position: relative;
-  border: 1px solid ${(props) => props.theme.color.gray300};
-  border-radius: 1rem;
-  padding: 0.5rem 1rem;
-  a {
-    text-decoration: none;
-    color: ${(props) => props.theme.color.fontColorBlack};
-  }
-  h3 {
-    ${Headline3};
-    padding: 0;
-    margin: 0;
-    ${({ theme }) => theme.device.mobile} {
-      font-size: 2.4rem;
-    }
-  }
-  .button-container {
-    display: flex;
-    justify-content: space-between;
-  }
-  .various-button-box {
-    button {
-      cursor: pointer;
-      border: 0;
-      background-color: unset;
-      padding: 0;
-      font-size: 2rem;
-      color: ${(props) => props.theme.color.gray400};
-    }
-    .delete-button {
-      margin-left: 1.8rem;
-    }
-  }
-`;
-const ModalChildren = styled.div`
-  display: flex;
-  flex-direction: column;
-  button {
-    cursor: pointer;
-    font-size: 2rem;
-    border: 0;
-    padding: 0.8rem 1.3rem;
-    border-radius: 0.5rem;
-    color: ${(props) => props.theme.color.fontColorWhite};
-    &:not(:first-child) {
-      margin-left: 0.8rem;
-    }
-  }
-  .button-container {
-    display: flex;
-    justify-items: flex-end;
-    align-self: flex-end;
-  }
-  .cancel-button {
-    background-color: ${(props) => props.theme.color.gray400};
-  }
-  .confirm-button {
-    background-color: ${(props) => props.theme.color.error700};
-  }
-`;
+import {
+  AddStoreButton,
+  Header,
+  Item,
+  ModalChildren,
+  StoreList,
+  Wrapper
+} from "./styles";
 
 const AdminStoreList = () => {
   const [toggleState, setToggleState] = useState<boolean | undefined>(false);
