+++ src/Page/Admin/Store/index.ts
@@ -1,0 +1,3 @@
+export { default as AdminCreateStore } from "./AdminCreateStore";
+export { default as AdminStoreList } from "./AdminStoreList";
+export { default as AdminUpdateStore } from "./AdminUpdateStore";
