+++ src/Page/Admin/Store/interface.ts
@@ -1,0 +1,7 @@
+export interface IStoreFormProps {
+  name: string;
+  code?: string;
+  phone: string;
+  address: string;
+  addFail?: string;
+}
